Fix CNV-92870: fix the v1 storage logic
When upgrading to this version, the CRD storage version is changed from
`v1beta1` to `v1`. Then the operator is trying to replace the
HyperConverged CR object in ETCD, to be in `v1` API format. If the
webhook of the previous version is still running, then this will work.
But if not, the new webhook pod is getting into crash loop, because it's
trying to read the HyperConverged CR before the webhook server started,
and so the conversion webhook is not available, and this read operation
fails. At the same time, the operator is failing to read the
HyperConverged CR, because the conversion webhook is not available.
Even reading the HC CR in v1beta1 format, and manually convert it,
fails, because the conversion webhook is needed for the update operation
as well.

To fix it, we:
1. moving the v1 storage logic from the operator to the webhook, so the
   operator does not depend on the webhook anymore.
2. the new code in the logic is moved after the webhook server starts,
   to make sure the conversion webhook is available for this process.
